ObsidianVault/workflow_ui/requirements.txt
# PURPOSE: Dependencies for workflow GUI (Phase 3).
# Run from ObsidianVault: pip install -r workflow_ui/requirements.txt
flask>=2.3.0
flask-limiter>=3.5
pyyaml>=6.0
pytest>=7.0.0
gradio>=4.0.0
playwright
pytest-playwright
